% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{functions}
\alias{functions}
\title{Favorite R functions of Sebastian Loos}
\format{
A dataframe with 5 variables and 110 observations
\describe{
\item{function}{function name}
\item{package}{package name}
\item{description}{short description about what the function does}
\item{documentation}{url to the documentation page}
\item{application}{application fiel where the function can be used}
}
}
\usage{
functions
}
\description{
This Dataset includes data about the functions mostly used by Sebastian Loos
}
\keyword{datasets}
